requirements.txt
kafka-python==2.0.2
pyspark==3.2.4
pymongo==4.6.1
streamlit==1.29.0
altair==5.2.0
vaderSentiment==3.3.2
pandas